chore(VR): Apply manual changes
